requirements.txt
psycopg2-binary==2.9.10
pandas==2.2.3
sentence-transformers==3.3.1
tqdm==4.67.1
requests==2.32.3
beautifulsoup4==4.12.3
fastapi==0.115.6
uvicorn==0.34.0
